+++ analyzers/stable_diffusion_analyzer.py
@@ -63,7 +63,7 @@ class StableDiffusionAnalyzer(BaseAnalyzer):
             "name": "compute_dtype",
             "type": "select",
             "choices": ["FP16", "INT8"],
-            "default": "FP16",
+             "default": "INT8",
             "description": "Compute data type"
         }
     ]
@@ -77,7 +77,7 @@ class StableDiffusionAnalyzer(BaseAnalyzer):
         tinme_steps=1,
         w_bit=16,
         a_bit=16,
-        compute_dtype="FP16"
+        compute_dtype="INT8"
     ):
         """
         return is a dict with the following format:
